Add files via upload
The 4th assignment asked to use the new AI SAM (Segment Anything Model) from Meta using different prompting methods to improve the work done in the first assignment for the x-ray dataset and correctly segment lung pictures. We had to also to create bounding boxes starting from the ground truth and used them to evaluate the performance of SAM. And we lastly had to create an object detection model, train it and then test it always against the SAM reporting different performance measures such as Intersection over Union and F1 score.